style(feature/DomainLayer): change AddAccount and AccountModel modifier to public
