--- SwiftCleanBootcamp/Data/Data.h
@@ -1,18 +1,0 @@
-//
-//  Data.h
-//  Data
-//
-//  Created by Willian Guedes on 20/10/22.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for Data.
-FOUNDATION_EXPORT double DataVersionNumber;
-
-//! Project version string for Data.
-FOUNDATION_EXPORT const unsigned char DataVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <Data/PublicHeader.h>
-
-
