+++ SwiftCleanBootcamp/DataTests/RemoteAddAccount/Tests/RemoteAddAccountTests.swift
@@ -35,14 +35,13 @@ final class RemoteAddAccountTests: XCTestCase {
         sut.add()
         XCTAssertEqual(httpClientSpy.url, url)
     }
-}
-
-class HttpClientSpy {
-    var url: URL?
-}
-
-extension HttpClientSpy: HttpPostClient {
-    func post(url: URL) {
-        self.url = url
+    
+    func test_add_should_call_httpClient_with_correct_data() {
+        let url = URL(string: "https://www.google.com.br")!
+        let httpClientSpy = HttpClientSpy()
+        let sut = RemoteAddAccount(url: url, httpClient: httpClientSpy)
+        
+        sut.add()
+        XCTAssertEqual(httpClientSpy.url, url)
     }
 }
