+++ SwiftCleanBootcamp/Domain/AddAccount.swift
@@ -7,13 +7,13 @@
 
 import Foundation
 
-protocol AddAccount {
+public protocol AddAccount {
     func add(addAccountModel: AddAccountModel, completition: @escaping (Result<AccountModel, Error>) -> Void)
 }
 
-struct AddAccountModel {
-    var name: String
-    var email: String
-    var password: String
-    var passwordConfirmation: String
+public struct AddAccountModel {
+    public var name: String
+    public var email: String
+    public var password: String
+    public var passwordConfirmation: String
 }
